pyproject.toml
[tool.poetry]
name = "stbot"
version = "0.1.0"
description = ""
authors = ["Alexander Høier <ahoi@inspari.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.30.0"
openai = "^1.10.0"
tqdm = "^4.66.1"
langchain = "^0.1.3"
pymupdf = "^1.23.19"
chromadb = "^0.4.22"
stqdm = "^0.0.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
